Add GetMovieData service
Add a service that requests the movie data from TMdb API using
a movie ID.
Add a MovieController that receives the requests and formulate a
response based on services.
Add routes that directs the requests to the controllers. This
route is imported at the main file.
Add .env file for TMdb API key and other sensitive data. An
example for this file is present and named .env.example.